requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-redis==4.10.0
django-timezone-field==3.0
idna==2.8
importlib-metadata==0.23
kafka-python==1.4.6
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
redis==3.2.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.8.0
tweetpy==0.1
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
